put sendRumorState in hrumorstate package
